impl: more progress on BufferedResourceConsumer, SimulationDriver refactorings
This commit actually encompasses some larger changes. To begin with,
`SimulationDriver` is refactored to work in terms of integer seconds, rather
than fractional time. This reflects a larger commitment to the concept that
the simulation has a lower bound of 1 Hz. Using integer time simplifies the
logic around resource management. Concurrently with this change, testing
infrastructure is added to make testing `BufferedResourceConsumer` much more
ergonomic and accurate.

+++ mod/Core/Simulator/ISimulated.cs
@@ -7,9 +7,9 @@ public interface ISimulated {
   /// The maximum Δt that the current state is valid for, assuming no outside events cause an
   /// invalidation.
   /// </summary>
-  double RemainingValidDeltaT { get; }
+  ulong RemainingValidDeltaT { get; }
 
-  void Tick(double deltaT);
+  void Tick(ulong deltaT);
   void OnSynchronized();
 
   void OnStabilized();
